Made sure model names can be written using upper and lower case

+++ src/make.c
@@ -736,7 +736,7 @@ calign *Make_Calign(int n_otu, int crunch_len, int state_len, int init_len, char
       cdata->c_seq[j]->name      = (char *)mCalloc((int)(strlen(sp_names_in[j])+1),sizeof(char));
       strcpy(cdata->c_seq[j]->name,sp_names_in[j]);
       cdata->c_seq[j]->state     = (char *)mCalloc(crunch_len*state_len+1,sizeof(char));
-      cdata->c_seq[j]->d_state   = (int *)mCalloc(crunch_len*state_len,sizeof(int));
+      cdata->c_seq[j]->d_state   = (short int *)mCalloc(crunch_len*state_len,sizeof(short int));
       cdata->c_seq[j]->is_ambigu = (short int *)mCalloc(crunch_len,sizeof(short int));
     }
 
@@ -746,7 +746,7 @@ calign *Make_Calign(int n_otu, int crunch_len, int state_len, int init_len, char
       cdata->c_seq_rm[j]->name      = (char *)mCalloc((int)(strlen(sp_names_out[j])+1),sizeof(char));
       strcpy(cdata->c_seq_rm[j]->name,sp_names_out[j]);
       cdata->c_seq_rm[j]->state     = (char *)mCalloc(crunch_len*state_len+1,sizeof(char));
-      cdata->c_seq_rm[j]->d_state   = (int *)mCalloc(crunch_len*state_len,sizeof(int));
+      cdata->c_seq_rm[j]->d_state   = (short int *)mCalloc(crunch_len*state_len,sizeof(short int));
       cdata->c_seq_rm[j]->is_ambigu = (short int *)mCalloc(crunch_len,sizeof(short int));
     }
